merges base operations and derived classes together
